+++ flutes/fs.py
@@ -4,7 +4,8 @@ import pickle
 import platform
 import shutil
 import subprocess
-from typing import Optional
+from pathlib import Path
+from typing import Iterator, Optional
 
 from .log import log
 from .types import PathType
@@ -16,6 +17,7 @@ __all__ = [
     "remove_prefix",
     "copy_tree",
     "cache",
+    "scandir",
 ]
 
 if platform.system() == "Darwin":
@@ -74,7 +76,7 @@ def remove_prefix(s: str, prefix: str) -> str:
     return s[prefix_len:]
 
 
-def copy_tree(src: PathType, dst: PathType, overwrite: bool = False):
+def copy_tree(src: PathType, dst: PathType, overwrite: bool = False) -> None:
     r"""Copy contents of folder ``src`` to folder ``dst``. The ``dst`` folder can exist or whatever (looking at you,
     :py:func:`shutil.copytree`).
 
@@ -124,3 +126,20 @@ def cache(path: Optional[PathType], verbose: bool = True, name: Optional[str] = 
         return wrapped
 
     return decorator
+
+
+def scandir(path: PathType) -> Iterator[PathType]:
+    r"""Lazily iterate over all files and directories under a directory. The returned path is the absolute path of the
+    child file or directory, with the same type as :attr:`path` (:py:class:`pathlib.Path` or :py:class:`str`).
+
+    :param path: Path to the directory.
+    :return: An iterator over children paths.
+    """
+    if isinstance(path, Path):
+        with os.scandir(path) as it:
+            for entry in it:
+                yield Path(entry.path)
+    else:
+        with os.scandir(path) as it:
+            for entry in it:
+                yield entry.path